+++ website/client/index.js
@@ -5,6 +5,9 @@ import ReactDOM from 'react-dom';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
+// Add Christmas mood ↓
+// import './components/happyNewYear';
+
 ReactDOM.render(<App />, document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change