+++ website/docs/components/feedback/examples/feedback.js
@@ -4,7 +4,7 @@ import Input from '@semcore/input';
 import { Box, Flex } from '@semcore/flex-box';
 import Link from '@semcore/link';
 import Dropdown from '@semcore/dropdown';
-import ChatXS from '@semcore/icon/lib/Chat/xs';
+import ChatXS from '@semcore/icon/Chat/m';
 import Textarea from '@semcore/textarea';
 import { Text } from '@semcore/typography';
 
@@ -93,7 +93,7 @@ class Feedback extends React.PureComponent {
           You can also send us an email to <Link>backlink.audit@semrush.com</Link>
         </FeedbackForm.Notice>
         <FeedbackForm.Notice hidden={status !== 'failed'} theme="danger">
-          Your message hasn’t been sent.
+          Your message has not been sent.
         </FeedbackForm.Notice>
       </FeedbackForm>
     );